custom attributes for SelectOneView
I had originally created a SelectOneView that allows labels
and values to be specified in XML. So say that you had dog,
cat, fish displayed to the user, the view would return a
d, c, or f as the "value" to be saved in the DB. Very similar
to how ODK Collect selects work. I exposed this to allow it
to be specified in XML, so you could add attributes to your
SelectOneView tag like custom:labels="@array/labels".
Unfortunately, it looks like Robolectric can't handle this,
and custom attributes cause it to fail (Bug #1102). So, I
am going to have to pull this out and specify it only in
code. However, I am committing this so that I can use it for
a test case later if I choose to try and fix the bug in
Robolectric.