add hasDni key to DniOrNameView
Before this commit, DniOrNameView was only adding values to
the map for the DNI and name EditTexts. This left open the
possibility that someone might enter both DNI and name by
toggling the RadioButtons, in which case sending the map
wouldn't make it clear which they actually wanted. To remedy
this, this commit adds a hasDni key that also indicates which
RadioButton was selected. It adds tests and updates all the
affected Activities.