Break a bunch of tests mostly relating to views
Getting these tests to pass will require knowledge about
views and activities.

Rule 1: DO NOT LOOK AT THE DIFF. This would tell you how to
fix them, but it would teach you nothing. Further, if you
saw that it isn't obvious how the change would fix a test,
that says nothing more than that you REALLY need to figure
this stuff out. All the more reason not to cheat.

Rule 2: DO NOT EDIT ANY TESTS. All of the tests will pass by
editing only source files and XML files.  Do NOT edit the
tests, or it will defeat the purpose. All the tests have
previously been passing. To the best of my knowledge there is
no reason to edit the tests.

The unit tests can run straight on the computer. There is a
single integration test (androidTest according to Android
Studio). You'll need to have an emulator or a connected
device to run that.